+++ JsonToMidiConverter/Models/Song/JsonConverters/AccentuatedConverter.cs
@@ -3,9 +3,9 @@ using System.Text.Json.Serialization;
 
 namespace JsonToMidiConverter.Models.Song.JsonConverters;
 
-public class AccentuatedConverter : JsonConverter<double>
+public class AccentuatedConverter : JsonConverter<float>
 {
-    public override double Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+    public override float Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
         if (reader.TokenType == JsonTokenType.True || reader.TokenType == JsonTokenType.False)
         {
@@ -13,12 +13,12 @@ public class AccentuatedConverter : JsonConverter<double>
         }
         else if (reader.TokenType == JsonTokenType.Number)
         {
-            return reader.GetDouble();
+            return reader.GetSingle();
         }
 
         return 0;
     }
 
-    public override void Write(Utf8JsonWriter writer, double value, JsonSerializerOptions options)
+    public override void Write(Utf8JsonWriter writer, float value, JsonSerializerOptions options)
         => JsonSerializer.Serialize(writer, value, options);
 }
